+++ Algorithms/C++/reverse_string.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 void reverse_string(string &str)
 {
-    int size = str.size();
+    int size = (int)str.size();
 
     for (int i = 0; i < size / 2; i++)
     {
