+++ Algorithms/C++/second_largest.cpp
@@ -27,7 +27,7 @@ int secondlargest(vector<int> &arr, int size)
 int main()
 {
     vector<int> arr = {4, 5, 6, 7, 45, 8, 9, 8};
-    int size = arr.size();
+    int size = (int)arr.size();
 
     int ans = secondlargest(arr, size);
